+++ module/scripts/constants.js
@@ -1,0 +1,18 @@
+////////////////////////////////////////////////////////////////////////////////
+//                   _____           _        _                               //
+//                  |  __ \         | |      | |                              //
+//                  | |__) |_ _  ___| | _____| |_ ___                         //
+//                  |  ___/ _` |/ __| |/ / _ \ __/ __|                        //
+//                  | |  | (_| | (__|   <  __/ |_\__ \                        //
+//                  |_|   \__,_|\___|_|\_\___|\__|___/ LIBRARY                //
+//                                                       By ZotyDev           //
+////////////////////////////////////////////////////////////////////////////////
+// ? Here are all the constants used by Packets, all static values are and
+// ? should be here.
+export class Constants {
+    static ID = 'packets';
+    static NAME_FLAT = 'Packets';
+    static NAME = `📦 ${Constants.NAME_FLAT}`;
+    static SOCKET;
+    static D;
+}