+++ CHANGELOG.md
@@ -1,1 +1,7 @@
-# MODULE Changelog
+# 📦 Packets Changelog
+
+## Version 1.0.0
+**Initial release**
+- *Internal* - Created a `Utils` class containing many utility functions.
+- *Internal* - When first loading the module, all the required folders and files will be created as soon as possible by any player with enough permissions.
+- *Internal* - Every action that needs replication is executed by the first player with enough permissions.